add unit test to check internal instrument conversions
